1. split integration tests 2. underscored unused args 3. commented exported api 4. enabled custom domain auth_test cases 5. typo, unuse var fixes 6. pagination support and bugfixes
